requirements.txt
requests==2.31.0
Pillow==10.0.1
matplotlib==3.7.2
numpy==1.24.3
tkinter-tooltip==2.2.0
python-dotenv==1.0.0